ipc: convert TimePoint Ipc.Snd/RcvTimeout to TimePeriod in sys_ipc()
As noted in a comment, this is the wrong way of handling a TimePoint
RcvTimeout unless a send phase is absent. Also changed redo_ipc_send_half() to
emit a TimePeriod, which (being a different porcelain for the send phase) is
completely valid.

There's now also an assert that thread_sleep()'s parameter is a TimePeriod.
There's basically no reason why it'd ever accept a TimePoint, and most
callsites couldn't supply a base clock anyway.