lib: add policy mechanism to the slab allocator
The policy controls allocation of slab pages, i.e. address space and physical
memory. The first (and so far only) policy lets the caller reserve an address
range for slabs of an object with size 2**n, so that object addresses can be
converted to a small unsigned integer and back with some bit operations. To
this end there's also some macros and a safe id-to-pointer function that's
supposed to never return a pointer to an unmapped region.

This interface could become more generic if another kind of allocation policy
appears. Doesn't seem like it, however.

[This changeset contains a bunch of tests for the policy slab allocator. They
cover most of the implementation proper by virtue of TDD, but nothing else.]

Relatedly, the policy allocator needed a reserve_heap_range() in heap.c, so
one was added. This could be paired with a free_heap_range() and backed with a
rb-tree indexed <struct as_free> tracking ranges instead of single pages.
Until some repeated path starts freeing policy slabs, though, there's no use
case.