mapdb: skip unrelated child refs in deref_child()
This adds an @eff_range parameter, which is equivalent to that in
unmap_entry_in_group() -- i.e. the map_entry's ->range is acceptable.

This optimization appears to be a bit hit-and-miss, but it doesn't hurt the
1-fork benchmark and sometimes appears to put the 8-fork one to around the
same cycles/fork line as the lesser benchmarks. C'est la vie; at least it
isn't breaking anything.