mapdb: rewrite parent's childref in expand_entry()
This only happens if the expanded entry was the high half of the resulting
entry, i.e. exactly half the time.

Also, the parent reference was being altered the wrong way: instead of
clearing a bit in the INDEX field it would destroy the MISC bits, the GROUP
bits, and INDEX up to the n_pages bit.

[fixing this exposed a bunch of knock-on problems in mapdb which were
subsequently also repaired, but the changes had to be merged in reverse order
because there were so many. so there's one reason to commit working shit on
top of shit that don't work.]