mapdb: add parent size condition to can_merge()
Oops.

This comes up when when forkserv receives grants for roottask pages in reverse
order (causing shrapnel because of page IDs), and then re-maps those items to
the sender which would otherwise eventually coalesce them, violating the
invariant.

Add deref_parent() to that end, though it originated elsewhere in the same
branch.