test: sched: add lipc_chain_timeout
It tests the effect of the RcvTimeout parameter of Ipc/Lipc on Lipc chains.
Most of the things it's supposed to test are currently missing, and many other
things aren't noted in the test's comment[0], but this a start -- and it has
failing test points at its position in the change series, so that's OK for
now.

Also, the test should have two extra variables for use of Ipc/Lipc and various
mixes thereof, but there we go.

[0] e.g. timeout length's effect on the chain timeout, effect of the chain
    timeout itself, etc.