sched, ipc: add partner tracking in sys_lipc(), returns in sys_ipc()
The net effect is that for partners created in Lipc, a reply with Ipc will
transfer control without going through the scheduler. This has some impact for
performance, but it's not notable enough to measure properly. Scheduling will
be correct because of the pre-existing partner pointer, which is nice.

Replies to non-partners via Lipc are also accounted for, though that isn't
strictly speaking related to this at all. The `sched_chain_timeout' global in
sched.c is also redundant if invariant checks aren't compiled in, for the time
being.

This changeset has multiple subcomponents. For one, it unexports *ing_thread()
in favour of sched_ipc_handoff_*(), as predicted in an earlier changeset. This
makes no difference in performance on its own.

Secondly, it adds the thread->u0.partner field, which is known not to conflict
with other items in <struct thread>.u0, and uses it to implement
return-to-partner semantics in Ipc. This makes intra-space Lcall-to-ReplyWait
cases faster, though the benefit will come equally to Ipc once it participates
in partner scheduling as well.

Thirdly, it adds some basic scheduling queue invariants. These cover the
`partner' field in <struct thread>, and the `sched_chain_timeout' global
variable. (these invariants came from an ugly [therefore unpublished] Z spec,
which explains the mathsiness and the weird $partner: THREAD \pfun THREAD$
map; it's just an inverse of t->u0.partner so that a value of NULL means $t
\notin \dom partner$.)

Fourthly, it updates various comments around sched.c and ipc.c, most notably
the explicit postcondition of ipc_send_half().

This changeset also leaves some things undone: Ipc doesn't partner with the
recipient or schedule into the recipient unless replying to an existing
partner, implying that the previous message came from a local thread via Lipc.
This limits any bugs' impact enough to make the code "proof of concept"
quality at best, but on the other hand it lets us skip all sorts of weird
testbench-inhibiting problems.

Also, there's no attempt to remove partnered threads or their partners from
the scheduling queue. So that overhead is still around.