test: thread: add test for self-modifying ThreadControl
This wasn't obvious, so it wasn't tested, and therefore didn't work.

Presumably this was because two of the modifiables are the sort that would
cause the calling thread to exit in an unusual way (i.e. space and utcb_pos),
thread suicide was already being tested for, and trying to set utcb_pos for an
activated thread is defined a no-op.

The test should furthermore confirm that scheduling messages go to the right
thread where applicable. As those messages aren't being generated, this will
be TODO for a while yet.