rewrite heap & kip to use <ukernel/memdesc.h>
That's to say, to initialize the kernel heap based on the
bootloader-constructed KCP, and to rebuild a KIP from that.

kmain.c needed some adjustments to cope with the id_start..id_end range being
only what the kernel's binary allocates statically, rather than that and the
kernel early reservation.

Also, mbiloader's BootInfo generation was altered to output a spec-compliant
"size" field. It'll now count the sizes of all the BootRecs indicated, but not
the space in between.