user: sigma0: build user heap with mdb_query()
The function is used to find the process VM shape, which breaks mightily when
some microkernel puts a hole smack in the middle. Which mung shouldn't be
doing.

Conventional memory goes into the "pumpable" section, i.e. freelists as well
as the by-range tree. Other memory goes into just the latter, with the
`readonly' option set for L4_ReadOnlyBIOSRegion (non-portable).

Also, change get_free_page*() to return a L4_Fpage_t w/ rights. This puts the
readonly part in when smaller pages are pulled from the range tree, and passes
it on in the IPC loop without fuss. Untested, but should work.