test: ipc: prevent shrimp case in map_smaller_pages
It happens, sometimes, that the forkserv page allocator maps consecutive
physical pages to a client. If that client then sends that larger page over
into a space with smaller mappings already present, this ends up invoking the
mapdb shrimp case -- which, being unimplemented, causes a panic().

So this changeset is strictly a work-around for a missing part of the map-over
functionality.