mapdb: break large child entries all the way in distribute_children()
This maintains the "parents larger or equal to children" invariant. Previously
there was a bug where a p_ent was chosen as the split source, which wouldn't
generate the correct confetti in the child due to either alignment or size.

The current approach finds the split-point, i.e. where the smallest fpages are
in the range before it gets larger again, and then works backward and forward
to use parent entries in ascending size order, which ensures that child
entries for fragments are never added to small parents (per the new assert in
split_entry()).