test: space: add mapdb_shrimps to provoke panic in ..._add_map()
This test used to provoke the old panic in that clause. It was subsequently
removed. There are some further variables for testing, but those aren't too
relevant until the faster single-pass algorithm comes around -- and since it
might not (for a mapdb format rework, say) we'll spare a bit of work here.