mapdb: write add_map shrimp case
This one is based on repeating calls to unmap_entry_in_group(). It wastes
cycles by moving potentially quite a few entries backward in the map_group,
but like all weird mapdb operations it's called too rarely for this to matter
at all.

However, a single-pass solution would be preferable just for simpler
termination guarantees, and for not entailing a backward goto.