test: ipc: avoid deadlock in IpcHelper::handle_fault
IpcHelper is used as a delay pager in point_xfer_timeouts. This means that
when a transfer timeout happens in its client, that client would resume
execution before the fault handler had replied.

If this let the client fault again while holding the dlmalloc spinlock, the
helper would eventually try to reply to the previous fault and fail (the
client being in passive send at that time), and then (due to the way muidl's
dispatchers work) end up calling free() -- causing a classic application-level
deadlock.

The fix makes IpcHelper::handle_fault() propagate to its pager on the client's
behalf, and then bow out of the IPC flow. This should eliminate the deadlock
from point_xfer_timeouts, and other tests like it.

Also, munge_case() was changed to use talloc to delay releasing its buffers,
so that its calls to free() can't interfere with measurement even if it were
significant, which the rest of the patch makes impossible anyhow.