ptab: clear TLB in pt_probe() access query
It turns out that core2 stops updating the access and dirty bits once it's got
a TLB entry where those bits are set. To clear them, and thereby get fresh
bits again, INVLPG is required.

INVLPG is only used when the iterator's space is the current space. This
should avoid flushing TLBs for access queries on non-current page tables.