remove <ukernel/guard.h>
The last time this has been useful was a couple of years ago. So it's
basically useless now.

In any case it can be resurrected when necessary; the idea and implementation
aren't all that bad -- just that nobody likes to keep a bunch of
GUARD_MEMBER()s around in their structs "just in case".