rework interrupt delivery and asynchronous preëmption
Unfortunately this is a consolidated changeset. Fortunately its major changes
are relatively few. So here they are.

Interrupt delivery was rewritten to allow kernel subsystems to set custom
bottom halves using an utility function. This required speccing out a protocol
for such handler functions, which is described in the inline documentation for
irq_handler_fn in <ukernel/interrupt.h>. In a nutshell, it allows each handler
to indicate that the current thread should resume execution, that the current
thread has been put to sleep and the kernel should invoke the scheduler, or
that a second thread has woken up and may preëmpt the currently-running
thread.

That protocol allows for a general method of handling asynchronous
preëmptions, i.e. those that occur due to a thread's IPC/xfer timeout, a
hardware interrupt, or (in the future) an IPC transaction from a thread
running on another CPU. The underlying resolution mechanism is also used to
handle preëmptions-in-potentia before kernel exit, to adequately process
interrupts that occurred while in kernel (i.e. deferred bottom halves); it'll
be used for resolving all kinds of synchronous preëmption in the future as
well.

Handling of interrupts reported to userspace was unified with in-kernel
interrupt handling, allowing for low-latency interrupt response in userspace
by simply scheduling interrupt-handling threads at a high priority. This will
allow an asynchronous scheduling mechanism based on multiple callbacks rather
than a single "on_preempt()" that reads a small amount of global state each
time.

Timer-based asynchronous preëmption handling was rewritten in terms of this
same interface. Its core is now implemented in sched.c:on_preempt(), which may
feed distinct kinds of exit porcelain for syscalls, interrupts, and
exceptions. Said porcelain replaces the previous return_from_exn(), the
saturday night special for irq0, and whatever schedule() used to have.

IPC transactions caused by asynchronous scheduling were reimplemented in, or
relocated to, sched.c. There's three of them: preëmption faults, delay
exceptions, and total_quantum=0 faults; each has a different effect on the
IPC/sched state machine.

Also, most of the preëmption-related test breakage introduced in the
refactored tests, earlier, was fixed in this changeset.

Minor changes:
  - ioapic_mask_irq() now requires disabled interrupts so that an interrupt
    tophalf doesn't send an EOI in between APIC accesses.
  - the new arbitrary interrupt bottom-half spec requires those functions to
    explicitly unmask a vector when the tophalf has masked it due to the
    interrupt repeating before the handler could be called.
  - tracing was added to irq.c under TRID_IRQ.
  - the distinct interrupt stack was removed. interrupt frames are now
    constructed on the ordinary syscall stack, making changes below a thread's
    %esp hazardous while interrupts are enabled.