sched: don't send tq=0 message when leaving threads in IPC
Instead set tq to 1 so that the message will be sent when the thread is next
descheduled.

Really entering_thread(), or its callers, should catch this case, and then
immediately call leaving_thread() & schedule another thread instead. This is
currently not happening, which lets certain threads enjoy an extra tick's
worth of total_quantum time.