sched: send preempt faults as the oneway clock-bearing form
And not as exceptions that must be replied to. The spec gives conflicting
information on when the preempt fault should be signaled, so until another
approach is judged better, mung will send preempt faults at resume and
preemption exceptions (i.e. the ones that occur when max_delay is exceeded)
at preempt. This should be noted in doc/scheduling.md .

This changeset mostly changes the way ipc.c handles oneway kernel-generated
IPC. To wit, it makes conditional the user-visible Ipc result stomp in
ipc_send_half(), straightens out some turns in ..._complete_oneway(), and
removes the option to disable the ipc_user() fastpath.

It also changes the todo_start() in sched:delay_yield and delay_exception as
it's now the preempt exception that isn't being generated correctly; and
reduces TODO coverage in preempt_exn_test to match test points that now pass.