sched: rename sched_next_thread() to enter_next_thread()
There was already a schedule_next_thread(), which finds the next thread that's
going to run for schedule(). Needless to say, that's quite a different thing
from what enter_next_thread() does. Oops.

Also, updated enter_next_thread() to possibly enter a preëmpting thread so as
to handle clock ticks while in kernel. This might be better off as a big huge
BUG_ON(), but since it's a good idea to deal with timer instability robustly
anyway, and entering_thread() already modifies *@next_p, let's leave it in.