test: sched: set spinner_fn()'s ExceptionHandler explicitly
The convention where a thread wrapper function sets ExceptionHandler to the
parent thread is about to be changed so that it's the ThreadMgr instance
instead. This'd break pump_wakeups(), but with this change it no longer will.

Also marked delay_without_signal's actual tests TODO because, somehow, this
changeset makes them unstable.