test: sched: add clear_preempt_flags_while_pending
What it says on the tin.

There's an extra ThreadSwitch in the test that shouldn't be there, but
regardless it's required to make the test work properly. The problem is
self-modifying Schedule, which is invisible to preëmption right now. The test
would work without this in the first iteration due to residual scheduling
effects from setup; with this ThreadSwitch it works every time. It should be
removed once tests exist for self-modifying Schedule's correct operation, and
the syscall has been corrected.