thread: return -EEXIST from thread_set_utcb() when requested
This indicates reuse of UTCB slots, which not all callers might permit. By my
reading, the L4.X2 spec allows the microkernel to forbid these; this doesn't
reduce the responsiblity of userspace not to try doing that.

The boolean parameter is there because eventually mung might permit sharing of
UTCB slots as an API extension.