testbench: add basic mutex conflict handling
The microkernel provides serialization as IPC to a dedicated thread which
executes at most one mutex operation at a time. Use of this thread is mandated
by setting a conflict bit in the mutex control word, which only the serializer
thread may clear.

There's also a bit of D_M_H tracing stuff, for help in dealing with deadlocks.