testbench: rewrite test ID generation
Now it's correct, and also slightly less painful. And slower, just in case our
friends in 1996 care, with their pentiums and their rap music.

This'll also make the testbench reporting script fail hilariously if it ever
enters the "re-run tests a small handful at a time" mode. Which is poorly
defined and without proper testing, so it might as well go away. But, the real
solution is to not generate test IDs redundantly but rather feed the ones that
testbench generates (see test.c:add_test_ids()) to the report script and pass
them along without regard for their actual contents.