test: self: add stubs for two thrd_*() routines & two tests for them
These stubs are for a future <threads.h>, and thrd_set_daemon_NP() is a local
runtime-specific extension.

The tests check whether setting the daemon bit allows a process to exit when
the other thread exits; and whether setting the daemon bit on the last
remaining thread causes the process to exit immediately.