testbench: add mutex coverage to thread.c global data
About time, too. This is easily the thorniest part of the multithreaded
runtime because of various interdependencies.

Also, thread.c is getting a bit long in the tooth. It should be reworked so
that all serializing operations happen in the manager thread, eliminating the
mutex stuff altogether. However, this seems to work for now, so let's not
break it for no good reason.