lib: rework <ukernel/invariant.h>
It should now be fast enough to not interfere with testbench execution, which
makes it useful for more than just targeted "harder" debugging.

The new interface also allows substructural invariant checks to use the
outermost INV_CTX. However, it doesn't work well with checks that're used both
from callers that have an INV_CTX, and ones that don't. There should likely be
some kind of INV_CTX_PARAM macro or something like that, but as mapdb.c didn't
require this it's not here.

Also unfuck space.c a little bit.