test: space: add huge_fanout
This verifies that massive reuse of a single page is possible. The case would
appear in e.g. a memory server's view of the most used part of a C runtime
library, which puts a copy of the mapping in every process in the system aside
from statically linked binaries.