test: space: add overmap_empty
This tests behaviour that's kind-of sort-of consistent with a reading of the
L4.X2 spec; in detail, that holes mapped from a source will unmap the
destination, leaving a hole. Previously it'd leave the old page behind.

Because the spec is ambiguous about this case, the test checks a feature
string and skips if it isn't present. Stuff was added for feature string
checks.