test: thread: don't require re-breath-of-life in err_on_lost_peer
Now it'll start the lost peer within the same process, rather than forking for
comparatively little reason. This is required because breath-of-life is
supposed not to happen when an active thread's version bits are altered.

This changeset also corrects two minor implementation flaws previously hidden
by forking (most likely forkserv's acting as a sync point). To wit, the new
version now takes a short nap to ensure "oth" gets to its receive phase, and
passes the parent TID to the child to not receive its initialization message
out of sync.