disable non-bitmap I/O in userspace
This was left enabled by erroneous reading of how x86 IOPL works. Also
recognize cli/sti in the I/O exception bottomhalf.

The necessary I/O fault passing has been added to testbench and forkserv,
which were also modified to recognize #UD from the kernel and signal it like
it were an ordinary segfault.

sigma0 was changed to report I/O faults if PF_TRACING is set to 1 in the
relevant #define in its source.