mapdb: fix probe_add() logic
It hadn't been updated for the world of buckets, so when called from
grow_children_array() (i.e. with @shares == NULL) it'd stop at the first
tombstone even while it was possible for there to be a bucket for @parent_ix
somewhere later down the line. With these changes, it crawls the chain to the
end to prove that there's no bucket there, or recognizes a bucket correctly
even when one occurs after a tombstone.

probe_add() could still be optimized by making it stop early when both a
tombstone and a non-bucket ref for @parent_ix have been seen. It's uncertain
whether the extra logic is worth it.

Also, fix a typo in add_child() to the effect that buckets are now recognized
by AUX_UP(...) == 6, not 7. There was even a macro for this.