user: remove unnecessary spinlock from syscalls-32.S
It'd be pretty rare to have multiple threads of control enter a syscall stub
simultaneously, because it's rare to have multiple threads of control in a
process before the first thread has hit a syscall. So let's remove this for
simplicity's sake, and because it appears to make something broken elsewhere
not quite as broken anymore.