ipc: add logic for closed_propagation
This also extends `flags' in <struct thread> to add TF_FROM_TURN for
indicating that the thread's FromSpec has been altered by propagation, so that
when it gets replied to its ActualSender will indicate where the message
really came from. If the FromSpec wasn't turned, ActualSender will be
unaltered; this allows e.g. IDL callers to clear the register and examine it
afterward to determine if the request was propagated by the immediate
recipient.

The L4.X2 spec says the "p" flag in MsgTag should only be set in that case if
the message was itself propagated, and we'll follow that spec; so unpropagated
replies to a turned FromSpec will have that flag cleared.

There may be another way to achieve this in a way that doesn't require IDL
compilers to provide a version of stubs that return the Virtual-/ActualSender
TCR, or the TF_FROM_TURN flag. If that shows up later, the flag (and the field
extension) may be removed.