testbench: load forkserv in non-idempotent memory
It doesn't need idempotent mappings, so it doesn't get any. This is required
when the microkernel reserves memory outside the KCP MemoryDescs without
parsing module ELF headers.

forkserv's heap is also now adjustable, i.e. it can start at the first 16M, or
at 32M, or so on until 256M where it stops for safety.