lib: shrink kmem_test to pass under large memory
It doesn't need to test a hundred times over, and worse, because the test is
subject to allocation vagaries, the recycling part isn't guaranteed to pass
anyway as it is right now.

More properly it should be changed to test that at least one of the pages
shows up after release, and not require all of them to do so.