testbench: add TRY_INIT to tap.c entrypoints
The general reason to add these is that if initialization happens due to (say)
a call within a subtest, such as when the plan is declared at end of test,
module initialization within the subtest plan_tests() call would run
tap_reset(), blowing away the subtest stack and ultimately aborting in the
corresponding subtest_end(). This applies to various other entrypoints besides
subtest_start() which also got a TRY_INIT each.

Also, don't inherit the outer TODO into subtests because it'd be confusing to
have seemingly mismatched todo_end() calls in a subtest even if it's only
during development.