lib: rework stdio
This permits separate stderr output from the microkernel, so that critical
messages like assert failures etc. from within the console output code itself
don't wind up in an endlessly recursive stack-trashing triple-faulting
deathapalooza.

Below the hood, this also renames fake_stdio.c to not be as fake since it's
started looking rather more proper recently, and adds line buffering and a
somewhat slower fwrite(3) which is nonetheless about as correct wrt flushing
as fputc(3) is, and that's what we want to see.