testbench: add fork_subtest_start/_end macros
These are very convenient for having off-process helpers that should produce
test data themselves. This adds subtest message stashing over the version
found in sneks, which the latter should imitate without resorting to POSIX
pipes and/or other complicating factors.

This also adds EINTR to <errno.h> because the macros imported from sneks
assume that's possible. It isn't, but the usage is more correct this way and
it's also simpler for when that part is copied back and forth again.

Also, tests. They pass immediately so they're in the same changeset.